Updated all template project.json tools to use 1.1
There were still tools dependencies in the project.json file that were depending on .Net Core 1.0.

This means that on a machine with only .Net Core 1.1 installed (like the microsoft/aspnetcore-build docker image) the `dotnet publish-iis` command would fail because it was trying to find .Net Core 1.0.

Using the latest 1.1.0-preview4-final versions of

- Microsoft.AspNetCore.Razor.Tools
- Microsoft.AspNetCore.Server.IISIntegration.Tools
- Microsoft.DotNet.Watcher.Tools

fixes the issue and makes the all publish work correctly on the docker machine.